rssde/requirements.txt
beautifulsoup4>=4.9.0
deluge-client>=1.9.0
feedparser>=6.0.0
PyYAML>=5.4.0
requests>=2.25.0
